# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AlterarAtivo": {
        "properties": {
          "ativo": {
            "title": "Ativo",
            "type": "boolean"
          }
        },
        "required": [
          "ativo"
        ],
        "title": "AlterarAtivo",
        "type": "object"
      },
      "AlterarPapel": {
        "properties": {
          "role": {
            "description": "admin ou staff",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "AlterarPapel",
        "type": "object"
      },
...
      "DefinirSenha": {
        "properties": {
          "nova_senha": {
            "title": "Nova Senha",
            "type": "string"
          }
        },
        "required": [
          "nova_senha"
        ],
        "title": "DefinirSenha",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TrocarSenha": {
        "properties": {
          "nova_senha": {
            "title": "Nova Senha",
            "type": "string"
          },
          "senha_atual": {
            "minLength": 1,
            "title": "Senha Atual",
            "type": "string"
          }
        },
        "required": [
          "senha_atual",
          "nova_senha"
        ],
        "title": "TrocarSenha",
        "type": "object"
      },
      "UnidadeCreate": {
        "properties": {
          "descricao": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "nome": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Nome",
            "type": "string"
          }
        },
        "required": [
          "nome"
        ],
        "title": "UnidadeCreate",
        "type": "object"
      },
      "UnidadesDoUsuario": {
        "properties": {
          "unidades": {
            "items": {
              "type": "integer"
            },
            "title": "Unidades",
            "type": "array"
          }
        },
        "title": "UnidadesDoUsuario",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Monitor de Alertas UPP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "api_register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Register Api Auth Register Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/unidades": {
      "get": {
        "operationId": "listar_unidades_endpoint_api_unidades_get",
        "parameters": [
          {
            "in": "query",
            "name": "incluir_inativas",
            "required": false,
            "schema": {
              "default": false,
              "title": "Incluir Inativas",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Listar Unidades Endpoint Api Unidades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar Unidades Endpoint",
        "tags": [
          "usuarios"
        ]
      },
      "post": {
        "operationId": "criar_unidade_endpoint_api_unidades_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnidadeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Criar Unidade Endpoint Api Unidades Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Criar Unidade Endpoint",
        "tags": [
          "usuarios"
        ]
      }
    },
...
    "/api/usuarios/eu/senha": {
      "post": {
        "description": "Troca da propria senha, exigindo a senha atual.\n\nExigir a atual impede que alguem com uma sessao aberta esquecida assuma a\nconta trocando a credencial.",
        "operationId": "trocar_propria_senha_api_usuarios_eu_senha_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrocarSenha"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trocar Propria Senha Api Usuarios Eu Senha Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trocar Propria Senha",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/usuarios/{username}/ativo": {
      "patch": {
        "operationId": "alterar_ativo_api_usuarios__username__ativo_patch",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlterarAtivo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Alterar Ativo Api Usuarios  Username  Ativo Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alterar Ativo",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/usuarios/{username}/papel": {
      "patch": {
        "operationId": "alterar_papel_api_usuarios__username__papel_patch",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlterarPapel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Alterar Papel Api Usuarios  Username  Papel Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alterar Papel",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/usuarios/{username}/senha": {
      "post": {
        "description": "Reset de senha por administrador (esquecimento, comprometimento).",
        "operationId": "definir_senha_de_usuario_api_usuarios__username__senha_post",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefinirSenha"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Definir Senha De Usuario Api Usuarios  Username  Senha Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Definir Senha De Usuario",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/usuarios/{username}/unidades": {
      "get": {
        "operationId": "unidades_do_usuario_endpoint_api_usuarios__username__unidades_get",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Unidades Do Usuario Endpoint Api Usuarios  Username  Unidades Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unidades Do Usuario Endpoint",
        "tags": [
          "usuarios"
        ]
      },
      "put": {
        "description": "Define quais unidades o usuario enxerga.\n\nLista vazia e valido e significa \"nao ve nada\" — deny by default, como em\n`unidades_do_usuario`. Nao e o mesmo que admin: admin ve tudo por causa do\nPAPEL, e esta lista nao o restringe.",
        "operationId": "definir_unidades_endpoint_api_usuarios__username__unidades_put",
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "title": "Username",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnidadesDoUsuario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Definir Unidades Endpoint Api Usuarios  Username  Unidades Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Definir Unidades Endpoint",
        "tags": [
          "usuarios"
        ]
      }
    },
...
  }
}